requirements.txt
alembic==1.13.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
WTForms==3.1.2
